+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change log
 
+-Simple Stack Extensions 2.2.0 (2021-03-08)
+--------------------------------
+- Added `core-ktx` module with the following new extensions:
+
+  - `Backstack.addRetainedObject()`
+  - `Backstack.removeRetainedObject()`
+  - `Backstack.hasRetainedObject()`
+  - `Backstack.getRetainedObject()`
+  - `Backstack.getRetainedObjectOrNull()`
+
+Easiest to use with: `import com.zhuinden.simplestackextensions.corektx.*`
+
+Core-ktx depends on features added in `simple-stack 2.6.0`, so `simple-stack-extensions 2.2.0` requires `simple-stack 2.6.0`.
+
+- Make generics have bound as `: Any` (instead of implicit `: Any?`).
+
+- Update Kotlin to 1.4.30.
+
+- Update to simple-stack 2.6.0.
+
 -Simple Stack Extensions 2.1.0 (2020-12-18)
 --------------------------------
 - Added `GlobalServices.get()` and `GlobalServices.getOrNull()` to `services-ktx`.